create shared code
