+++ src/frontend/admin/src/App.js
@@ -15,7 +15,8 @@ import BackendLog from "./components/BackendLog";
 import Dashboard from "./components/Dashboard";
 import Analytics from "./components/Analytics";
 import TeamSpotlight from "./components/TeamSpotlight";
-import { createApiGet, setFavicon, isShouldUseDarkColor, useLocalStorageState } from "./utils";
+import { createApiGet } from "shared-code/utils";
+import { setFavicon, isShouldUseDarkColor, useLocalStorageState } from "./utils";
 import FullScreenClockManager from "./components/FullScreenClockManager";
 import AdvancedJson from "./components/AdvancedJson";
 import MediaFiles from "./components/MediaFiles";
