+++ src/frontend/admin/src/components/Advertisement.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Container from "@mui/material/Container";
 import { useSnackbar } from "notistack";
-import { errorHandlerWithSnackbar } from "../errors";
+import { errorHandlerWithSnackbar } from "shared-code/errors";
 import { PresetsManager } from "./PresetsManager";
 import { usePresetWidgetService } from "../services/presetWidget";
