+++ src/frontend/common/src/errors.js
