--- src/frontend/locator/src/styles.js
@@ -1,15 +1,0 @@
-export const faviconTemplate = `<svg width="512" height="512" viewBox="0 0 512 512" fill="none" xmlns="http://www.w3.org/2000/svg">
-<g clip-path="url(#clip0_1734_27177)">
-<rect width="512" height="512" rx="256" fill="{CONTEST_COLOR}"/>
-<circle cx="255.495" cy="256.494" r="42.75" fill="{TEXT_COLOR}"/>
-<path d="M63.125 256.5C63.125 150.254 149.254 64.125 255.5 64.125" stroke="{TEXT_COLOR}" stroke-width="48.0938" stroke-linecap="round"/>
-<path d="M447.875 256.5C447.875 362.746 361.746 448.875 255.5 448.875" stroke="{TEXT_COLOR}" stroke-width="48.0938" stroke-linecap="round"/>
-<path d="M148.62 256.494C148.62 197.469 196.469 149.619 255.495 149.619" stroke="{TEXT_COLOR}" stroke-width="48.0938" stroke-linecap="round"/>
-<path d="M362.381 256.506C362.381 315.531 314.531 363.381 255.506 363.381" stroke="{TEXT_COLOR}" stroke-width="48.0938" stroke-linecap="round"/>
-</g>
-<defs>
-<clipPath id="clip0_1734_27177">
-<rect width="512" height="512" rx="256" fill="{TEXT_COLOR}"/>
-</clipPath>
-</defs>
-</svg>`;
